Atmosphere

A dummy dataset containing minimum atmospheric data

Usage

atmosphere

Format

A data frame with 12 rows and 7 variables:

tAtm

Time for which the i-th data record is provided (T)

Prec

Precipitation rate (LT-1) (in absolute value)

rSoil

Potential evaporation rate (LT-1) (in absolute value). rSoil(i) is interpreted as KodTop when a time variable Dirichlet or Neumann boundary condition is specified.

rRoot

Potential transpiration rate (LT-1) (in absolute value).

hCritA

Absolute value of the minimum allowed pressure head at the soil surface (L).

Source

https://raw.githubusercontent.com/phydrus/phydrus/dev/examples/phydrus_paper/data/atmosphere.csv/

Examples

head(atmosphere)
#>   tAtm Prec      rSoil      rRoot hCritA
#> 1    1 0.51 0.00000000 0.01464508 100000
#> 2    2 0.02 0.00000000 0.01842473 100000
#> 3    3 0.00 0.03394404 0.03132639 100000
#> 4    4 0.00 0.00000000 0.02748290 100000
#> 5    5 0.00 0.03357781 0.02251188 100000
#> 6    6 0.00 0.03713137 0.02779009 100000