Prepare Atmosphere

Prepares atmospheric input data structure required by HYDRUS1D and by default uses a conservative tracer in irrigation source "clearwater" ( set to 1) in order to track the share of cleaned wastewater in the system inflow rate (as "Prec" column is a combined value of irrigation using either "groundwater" or "clearwater" and real "rainfall").

Usage

prepare_atmosphere(
  atm,
  conc_irrig_clearwater = 100,
  conc_irrig_groundwater = 0,
  conc_rain = 0,
  defaults = kwb.hydrus1d::defaults_atmosphere()
)

Arguments

atm

atm as retrieved by prepare_atmosphere_data

conc_irrig_clearwater

substance concentration in source "clearwater" used for irrigation (default: 100, set all other source concentrations in default to 0 in order to calculate share of "clearwater" infiltration to groundwater)

conc_irrig_groundwater

substance concentration in source "groundwater" used for irrigation (default: 0)

conc_rain

substance concentration in rainfall (default: 0)

defaults

defaults for undefined parameterskwb.hydrus1d::defaults_atmosphere()

Value

tibble with peoered

Examples

atm <- prepare_atmosphere_data()
atm_selected <- select_hydrologic_years(atm)
prepare_atmosphere(atm_selected)
#> # A tibble: 1,279 × 14
#>     tAtm   Prec rSoil rRoot hCritA    rB    hB    ht  tTop  tBot  Ampl  cTop
#>    <int>  <dbl> <dbl> <int>  <int> <int> <int> <int> <int> <int> <int> <dbl>
#>  1     1 0.0645 0.310     0 100000     0     0     0     0     0     0 100. 
#>  2     2 0.615  0.155     0 100000     0     0     0     0     0     0  10.5
#>  3     3 0.0645 0.240     0 100000     0     0     0     0     0     0 100. 
#>  4     4 0.245  0.12      0 100000     0     0     0     0     0     0  26.4
#>  5     5 0.0745 0.08      0 100000     0     0     0     0     0     0  86.6
#>  6     6 0.0645 0.19      0 100000     0     0     0     0     0     0 100. 
#>  7     7 0.0645 0.284     0 100000     0     0     0     0     0     0 100. 
#>  8     8 0.105  0.148     0 100000     0     0     0     0     0     0  61.7
#>  9     9 0.0645 0.24      0 100000     0     0     0     0     0     0 100. 
#> 10    10 0.0645 0.24      0 100000     0     0     0     0     0     0 100. 
#> # … with 1,269 more rows, and 2 more variables: cBot <int>, RootDepth <int>