Prepare Atmosphere

Prepares atmospheric input data structure required by HYDRUS1D and by default uses a conservative tracer in irrigation source "clearwater" ( set to 1) in order to track the share of cleaned wastewater in the system inflow rate (as "Prec" column is a combined value of irrigation using either "groundwater" or "clearwater" and real "rainfall").

Usage

prepare_atmosphere(
  atm,
  conc_irrig_clearwater = 100,
  conc_irrig_groundwater = 0,
  conc_rain = 0,
  defaults = kwb.hydrus1d::defaults_atmosphere()
)

Arguments

atm

atm as retrieved by prepare_atmosphere_data

conc_irrig_clearwater

substance concentration in source "clearwater" used for irrigation (default: 100, set all other source concentrations in default to 0 in order to calculate share of "clearwater" infiltration to groundwater)

conc_irrig_groundwater

substance concentration in source "groundwater" used for irrigation (default: 0)

conc_rain

substance concentration in rainfall (default: 0)

defaults

defaults for undefined parameterskwb.hydrus1d::defaults_atmosphere()

Value

tibble with peoered

Examples

atm <- prepare_atmosphere_data()
atm_selected <- select_hydrologic_years(atm)
prepare_atmosphere(atm_selected)
#> # A tibble: 2,375 × 13
#>     tAtm  Prec rSoil rRoot hCritA    rB    hB    ht  tTop  tBot  Ampl  cTop
#>    <int> <dbl> <dbl> <int>  <int> <int> <int> <int> <int> <int> <int> <dbl>
#>  1     1 0.106 0.240     0 100000     0     0     0     0     0     0  99.3
#>  2     2 0.656 0.187     0 100000     0     0     0     0     0     0  16.1
#>  3     3 0.106 0.211     0 100000     0     0     0     0     0     0  99.3
#>  4     4 0.286 0.148     0 100000     0     0     0     0     0     0  36.8
#>  5     5 0.116 0.134     0 100000     0     0     0     0     0     0  90.7
#>  6     6 0.106 0.247     0 100000     0     0     0     0     0     0  99.3
#>  7     7 0.106 0.212     0 100000     0     0     0     0     0     0  99.3
#>  8     8 0.146 0.137     0 100000     0     0     0     0     0     0  72.1
#>  9     9 0.106 0.219     0 100000     0     0     0     0     0     0  99.3
#> 10    10 0.106 0.289     0 100000     0     0     0     0     0     0  99.3
#> # ℹ 2,365 more rows
#> # ℹ 1 more variable: cBot <dbl>